+++ src/io/fasta.rs
@@ -198,8 +198,7 @@ impl Index {
             .map(|record| Sequence {
                 name: record.name.clone(),
                 len: record.len,
-            })
-            .collect()
+            }).collect()
     }
 }