+++ src/io/gff.rs
@@ -27,7 +27,7 @@ use std::path::Path;
 
 use csv;
 
-use utils::Strand;
+use bio_types::strand::Strand;
 
 /// `GffType`
 ///
@@ -354,8 +354,8 @@ impl Record {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use bio_types::strand::Strand;
     use multimap::MultiMap;
-    use utils::Strand;
 
     const GFF_FILE: &'static [u8] = b"P0A7B8\tUniProtKB\tInitiator methionine\t1\t1\t.\t.\t.\t\
 Note=Removed,Obsolete;ID=test